requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
bcrypt==4.0.0
cffi==1.15.1
construct==2.10.68
cryptography==38.0.1
fire==0.4.0
future==0.18.2
lxml==4.9.1
paramiko==2.11.0
pycparser==2.21
pycryptodomex==3.15.0
pykeepass==4.0.3
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
six==1.16.0
termcolor==2.0.1
tornado==6.2
